requirements.txt
# Core DL
torch>=2.0.0
torchvision>=0.15.0

# Model hub (EVA-02, ConvNeXt V2, Swin Transformer via timm)
timm>=0.9.16

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.4.0

# Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization & XAI
matplotlib>=3.7.0
seaborn>=0.13.0

# Progress bars & utilities
tqdm>=4.66.0

# Optional: faster JPEG decoding (recommended for large datasets)
# Pillow-SIMD>=9.0.0

# Optional: for Colab / Kaggle GPU training
# accelerate>=0.25.0
